Add finished initial commands running in both modes
This commit added functionality related with running the same functions from bot
modes (interactive and non-interactive). Functions are now part of dictionary
which make allocation of command much more easier.

Also add variables to detect 'interactive' or 'not-interactive' mode.

Last change is adding very initial functionality for 'init' target
(so far only display message - but this is a start :))