Add initial version of importing boilr template
This commit adds basic implementation to importing boilr template (after
boilr installation via 'init' target). Templates are kept locally together with
IC4A projects. This allow add templates to boilr from the same local folder.
Benefit of this for this case that a lot of templates can be added very quickly.

One of the idea is using templates as best as possible during development of
this project so this will allow quickly create new items based on templates.